requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
requests==2.32.3
pillow==11.2.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
sqlparse==0.5.3
uritemplate==4.1.1
fakeredis==2.29.0
pymongo==4.12.1
celery==5.5.2
psycopg2-binary==2.9.10
django-celery-beat==2.8.0
gunicorn==23.0.0
python-decouple==3.8